refactor: remove old data change handler
